requirements.txt
django
djangorestframework
pandas
nltk
matplotlib
seaborn
wordcloud
numpy